requirements.txt
numpy
faiss-cpu
tqdm
PyPDF2
openai
python-dotenv
PyMuPDF